Update melgeek/mojo75 vid and pid Add    melgeek/mojo75/rev1/rev1.c
