+++ keyboards/melgeek/mojo75/rev1/config.h
@@ -17,8 +17,8 @@
 #pragma once
 
 /* USB Device descriptor parameter */
-#define VENDOR_ID       0xFEED
-#define PRODUCT_ID      0xED75
+#define VENDOR_ID       0xEDED
+#define PRODUCT_ID      0x7075
 #define DEVICE_VER      0x0001
 #define MANUFACTURER    MelGeek
 #define PRODUCT         MOJO75
